Add Table Style

Given a provided table style, add it to the tableStyle file in the document. this makes it available to tables that get added to the powerpoint as a table style option

add_table_style(
  pptx,
  table_style,
  verbose = getOption("mirage.verbose", default = FALSE),
  error_call = caller_env()
)

Arguments

pptx

A Powerpoint wrapped in a <pptx_container> object. see load_pptx()

table_style

an xml_node that defines a tblStyle. Created by new_table_style().

verbose

if TRUE, some information is cli::cli_inform()ed along the way

error_call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.