Add or remove content from a slide

content() lets a user specify a value to add to a slide, and the location on the slide to add it through the ph argument. ph defines the placeholder, or height, width, x and y offset of the bounding box where the content should be added.

content(
  value,
  ph = ph_body(),
  metadata = NULL,
  ...,
  group_contents = TRUE,
  error_call = current_env()
)

remove_content(ph, error_call = current_env())

Arguments

value

value to be polished and then added at the placeholder defined

ph

Defining where values should be added to or removed from. Must be a placeholder object

metadata

named character vector of metadata to store in the p:extLst/p:ext/custom:meta/custom:metadata nodes

...

arguments to include when polishing value

group_contents

boolean indicating if the polished value should be shown in the powerpoint as a grouped Sp. The default, and for most cases, TRUE.

error_call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

Details

remove_content() on the other hand, lets users specify a placeholder with content that should be removed from the slide using the same placeholder nomenclature. The placeholder must exist in the slide to be removed.

Examples

content("My Title 2", ph = ph_title())
#> $value
#> [1] "My Title 2"
#> 
#> $ph
#> ── reference placeholder with settings:  ───────────────────────────────────────
#> List of 6
#>  $ type       : chr [1:2] "title" "ctrTitle"
#>  $ label      : chr "[Tt]itle"
#>  $ label_from : chr [1:2] "layout" "slide"
#>  $ label_match: chr "match"
#>  $ tie_breaker: chr "top"
#>  $ replace    : logi TRUE
#> ── new placeholder with settings:  ─────────────────────────────────────────────
#> List of 6
#>  $ label: chr "mirage_title_placeholder"
#>  $ type : chr "title"
#>  $ offx : chr "5%"
#>  $ offy : chr "5%"
#>  $ cx   : chr "90%"
#>  $ cy   : chr "12.5%"
#> 
#> $polishing_args
#> list()
#> 
#> $metadata
#> NULL
#> 
#> $group_contents
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "mirage_content"
content(mtcars      , ph = ph_body())
#> $value
#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Mazda RX4           21.0   6 160.0 110 3.90 2.620 16.46  0  1    4    4
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Datsun 710          22.8   4 108.0  93 3.85 2.320 18.61  1  1    4    1
#> Hornet 4 Drive      21.4   6 258.0 110 3.08 3.215 19.44  1  0    3    1
#> Hornet Sportabout   18.7   8 360.0 175 3.15 3.440 17.02  0  0    3    2
#> Valiant             18.1   6 225.0 105 2.76 3.460 20.22  1  0    3    1
#> Duster 360          14.3   8 360.0 245 3.21 3.570 15.84  0  0    3    4
#> Merc 240D           24.4   4 146.7  62 3.69 3.190 20.00  1  0    4    2
#> Merc 230            22.8   4 140.8  95 3.92 3.150 22.90  1  0    4    2
#> Merc 280            19.2   6 167.6 123 3.92 3.440 18.30  1  0    4    4
#> Merc 280C           17.8   6 167.6 123 3.92 3.440 18.90  1  0    4    4
#> Merc 450SE          16.4   8 275.8 180 3.07 4.070 17.40  0  0    3    3
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Merc 450SLC         15.2   8 275.8 180 3.07 3.780 18.00  0  0    3    3
#> Cadillac Fleetwood  10.4   8 472.0 205 2.93 5.250 17.98  0  0    3    4
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4
#> Chrysler Imperial   14.7   8 440.0 230 3.23 5.345 17.42  0  0    3    4
#> Fiat 128            32.4   4  78.7  66 4.08 2.200 19.47  1  1    4    1
#> Honda Civic         30.4   4  75.7  52 4.93 1.615 18.52  1  1    4    2
#> Toyota Corolla      33.9   4  71.1  65 4.22 1.835 19.90  1  1    4    1
#> Toyota Corona       21.5   4 120.1  97 3.70 2.465 20.01  1  0    3    1
#> Dodge Challenger    15.5   8 318.0 150 2.76 3.520 16.87  0  0    3    2
#> AMC Javelin         15.2   8 304.0 150 3.15 3.435 17.30  0  0    3    2
#> Camaro Z28          13.3   8 350.0 245 3.73 3.840 15.41  0  0    3    4
#> Pontiac Firebird    19.2   8 400.0 175 3.08 3.845 17.05  0  0    3    2
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Porsche 914-2       26.0   4 120.3  91 4.43 2.140 16.70  0  1    5    2
#> Lotus Europa        30.4   4  95.1 113 3.77 1.513 16.90  1  1    5    2
#> Ford Pantera L      15.8   8 351.0 264 4.22 3.170 14.50  0  1    5    4
#> Ferrari Dino        19.7   6 145.0 175 3.62 2.770 15.50  0  1    5    6
#> Maserati Bora       15.0   8 301.0 335 3.54 3.570 14.60  0  1    5    8
#> Volvo 142E          21.4   4 121.0 109 4.11 2.780 18.60  1  1    4    2
#> 
#> $ph
#> ── reference placeholder with settings:  ───────────────────────────────────────
#> List of 6
#>  $ type       : chr [1:2] "body" "img"
#>  $ label      : chr "(([Cc]ontent)|([Tt]ext)) [Pp]laceholder"
#>  $ label_from : chr [1:2] "layout" "slide"
#>  $ label_match: chr "match"
#>  $ tie_breaker: chr "largest"
#>  $ replace    : logi TRUE
#> ── new placeholder with settings:  ─────────────────────────────────────────────
#> List of 6
#>  $ label: chr "mirage_body_placeholder"
#>  $ type : chr "body"
#>  $ offx : chr "5%"
#>  $ offy : chr "27.5%"
#>  $ cx   : chr "90%"
#>  $ cy   : chr "67.5%"
#> 
#> $polishing_args
#> list()
#> 
#> $metadata
#> NULL
#> 
#> $group_contents
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "mirage_content"

remove_content(ph = ph_title())
#> $value
#> NULL
#> 
#> $ph
#> ── reference placeholder with settings:  ───────────────────────────────────────
#> List of 6
#>  $ type       : chr [1:2] "title" "ctrTitle"
#>  $ label      : chr "[Tt]itle"
#>  $ label_from : chr [1:2] "layout" "slide"
#>  $ label_match: chr "match"
#>  $ tie_breaker: chr "top"
#>  $ replace    : logi TRUE
#> ── new placeholder with settings:  ─────────────────────────────────────────────
#> List of 6
#>  $ label: chr "mirage_title_placeholder"
#>  $ type : chr "title"
#>  $ offx : chr "5%"
#>  $ offy : chr "5%"
#>  $ cx   : chr "90%"
#>  $ cy   : chr "12.5%"
#> 
#> $polishing_args
#> $polishing_args$replace
#> [1] TRUE
#> 
#> 
#> $metadata
#> NULL
#> 
#> $group_contents
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "remove_mirage_content" "mirage_content"       
remove_content(ph = ref_placeholder(label = "An Existing Placeholder"))
#> $value
#> NULL
#> 
#> $ph
#> ── reference placeholder with settings:  ───────────────────────────────────────
#> List of 6
#>  $ type       : chr [1:9] "body" "title" "ctrTitle" "ftr" ...
#>  $ label      : chr "An Existing Placeholder"
#>  $ label_from : chr [1:2] "slide" "layout"
#>  $ label_match: chr "exact"
#>  $ tie_breaker: chr [1:6] "largest" "smallest" "left" "right" ...
#>  $ replace    : logi TRUE
#> 
#> $polishing_args
#> $polishing_args$replace
#> [1] TRUE
#> 
#> 
#> $metadata
#> NULL
#> 
#> $group_contents
#> [1] TRUE
#> 
#> attr(,"class")
#> [1] "remove_mirage_content" "mirage_content"