Estimate constant on-target mutation rates

estimate_inference_with_empirical assumes a constant on-target mutation or conversion rate for all sites regardless of sequence context, and estimates site-level signal using a maximum-likelihood approach.

Usage

estimate_inference_with_empirical(
  chr.info,
  treatment_fixed_count,
  control_fixed_count,
  treatment_depth,
  control_depth,
  delta = 0.001,
  depth.cutoff = 10,
  homo.cutoff = 0.99,
  lambda1 = "auto",
  top.sites = NULL,
  bg.method = "fisher",
  bg.target = "treatment",
  highly.methyl.cutoff = 0.95,
  seed = NULL,
  thread = 1
)

Arguments

chr.info

A data.frame with three columns: site position in the format "chrom_coordinate_strand" (e.g., chr1_14404_-), site motif sequence (e.g., AAACC), and site motif type (e.g., NNUNN).

treatment_fixed_count

A vector of non-mutation counts for each site in the treatment sample.

control_fixed_count

A vector of non-mutation counts for each site in the control sample.

treatment_depth

A vector of read coverage for each site in the treatment sample.

control_depth

A vector of read coverage for each site in the control sample.

delta

The sequencing error rate. Default is 0.001.

depth.cutoff

The minimum read coverage for each site in the treatment and control sample. Default is 10.

homo.cutoff

The minimum ratio of non-mutation counts for each site in the control sample that is used to determine homozygous sites. Default is 0.99.

lambda1

The method to estimate the on-target mutation rate. One of "auto", "site" or a numeric value between 0 and 1. If "auto", the on-target mutation rate is estimated from the upper quantile of candidate signal sites, where the quantile is defined by highly.methyl.cutoff. If "site", the on-target mutation rate is estimated based on a user-defined site set that is provided by the argument top.sites. If a numeric value between 0 and 1, the on-target mutation rate is explicitly defined as the value. Default is "auto".

top.sites

A site set used to estimate the on-target rate. One line or vector element should correspond to one site ID. The argument is only used if lambda1 = "site".

bg.method

The statistical test to infer the background sites. One of "binomial", "fisher", "lrt". Default is "fisher".

bg.target

A string indicating what type of read count data used to estimate the background mutation rate. One of "treatment", "control" or "both". If "treatment", uses the read count data from the treatment sample; if "control", from the control sample; if "both", from both. Default is "treatment".

highly.methyl.cutoff

The upper quantile cutoff of candidate signal sites used for estimating the on-target mutation rate. This argument name is retained for backward compatibility. Default is 0.95.

seed

A seed for the random number generators.

thread

A positive integer numeric specifying the number of cores for parallel computing. Default is 1.

Value

A list including a data.frame of homozygous sites, a data.frame of heterozygous sites, the estimated on-target mutation rate (lambda1), and the estimated background mutation rate (lambda2).

Details

estimate_inference_with_empirical takes a site-level count table as input. The table can be generated by any preprocessing pipeline as long as it contains site IDs, motif annotations, treatment and control fixed-base counts, and treatment and control read depths in the expected order. If one would like to estimate the on-target mutation rate using a user-defined site set, an extra site set is required. See the argument top.sites for details.

Examples

# Estimate signal levels from a PAR-CLIP count table.
count_table <- read.delim(system.file("extdata", "pos.read.count.example.parclip.txt",
                                      package = "MIRAGE"))
res <- with(count_table, estimate_inference_with_empirical(cbind(pos, motif, type),
            treated_fixed_count, control_fixed_count, treated_depth, control_depth,
            delta = 0.001, depth.cutoff = 10, lambda1 = "auto", top.sites = NULL,
            bg.method = "lrt", bg.target = "both", highly.methyl.cutoff = 0.95,
            seed = 123, thread = 1))
#> Considering 989 (99%) sites as homozygous...
#> Considering 794 (80.28%) sites as background (homozygous sites only)...
#> Considering 29 (2.93%) sites as high-signal site candidates (homozygous sites only)...
head(res$homosites[, c("pos", "motif", "beta_est", "lrt_fdr")])
#>              pos motif     beta_est   lrt_fdr
#> 1 chr1_1233912_+ ACUUU 1.006474e-06 1.0000000
#> 2 chr1_1233913_+ CUUUC 1.006474e-06 1.0000000
#> 3 chr1_1233914_+ UUUCC 1.161686e-01 0.4341611
#> 4 chr1_1233917_+ CCUUU 9.218868e-02 0.4656565
#> 5 chr1_1233918_+ CUUUC 2.120883e-01 0.3152392
#> 6 chr1_1233919_+ UUUCA 9.218868e-02 0.4656565