Polish fit models

Polish fit models

# S3 method for class 'preset_shape'
polish_content_pptx(x, ph = "<p:ph/>", pptx, ..., error_call = current_env())

Arguments

x

preset_shape object to polish

ph

placeholder

pptx

The Powerpoint Presentation object

...

These dots are for future extensions and must be empty.

error_call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

Examples

pptx <- example_pptx() |>
 add_slide(
  layout = "Title and Content",
  content(shape("rightArrow", fill = "#000000", outline = "#ffffff"), ph = ph_title()),
  content(shape("heart", fill = "#ffffff", outline = "#000000"), ph = ph_body()),
  content(
    shape("funnel"),
    ph = new_placeholder(x_offset = "10%",y_offset="20%", height = "20%",width="22%")
   )
 )
#> → mirage to add 3 contents at slide 4.
#> ✔ Successfully added content to placeholder "Title 1".
#> ✔ Successfully added content to placeholder "Content Placeholder 2".
#> ✔ Successfully added content to placeholder "new_body_placeholder_7".

file <- tempfile(fileext = ".pptx")
pptx |> save_pptx(file)